% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnndm.R
\name{bnndm}
\alias{bnndm}
\title{NN distance based block CV}
\usage{
bnndm(
  x = NULL,
  modeldomain = NULL,
  samplesize = 1000,
  sampling = "regular",
  n_steps = 10,
  maxn = 10,
  cl = NULL
)
}
\arguments{
\item{x}{A sf object containing the training points.}

\item{modeldomain}{A raster or terra object containing the predictor variables.}

\item{samplesize}{Optional. The number of samplepoints generated in the modeldomain to describe the prediction space.
If NULL (default), the samplesize equals the number of cells in the predictor raster divided by ten.}

\item{sampling}{The type in which the samplepoints are generated. The default value is "Fibonacci", which is recommended for global studies.}

\item{n_steps}{The number of steps until termination.}

\item{maxn}{The maximal number of clusters to be fitted. By default 10.}

\item{cl}{A cluster object.}
}
\description{
This function estimates the optimal cross-validation folds.
Therefore the function calculates the nearest neighbor distance
between training locations and prediction locations (Gij), as well as
between different cross-validation folds  (Gjstar). The Wasserstein-Test is used
to select the CV-folds yielding the lowest distance between Gjstar and Gij.
}
\details{
The nearest neighbor distance distribution between training and prediction locations,
as well as between cross-validation folds are calculated. Based on that, the
cross-validation split yielding the lowest distance between those distributions is chosen.

TBC.
}
\note{
Experimental.
}
\examples{
\dontrun{
library(sf)
library(raster)
library(ggplot2)

### prepare sample data:
dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
dat <- aggregate(dat[,c("DEM","TWI", "NDRE.M", "Easting", "Northing","VW")],
   by=list(as.character(dat$SOURCEID)),mean)
pts <- dat[,-1]
pts <- st_as_sf(pts,coords=c("Easting","Northing"))
st_crs(pts) <- 26911
studyArea <- raster::stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))

bnndm_folds <- bnndm(pts, modeldomain=studyArea)
bnndm_folds$stat # Integral between the two curves
plot(bnndm_folds)
plot_geodist(pts, studyArea, cvfolds = bnndm_folds$indx_test)
ggplot() +
  geom_sf(data = pts, col = bnndm_folds$clusters)

#use for cross-validation:
library(caret)
ctrl <- trainControl(method="cv",
   index=bnndm_folds$indx_train)
model_bnndm <- train(dat[,c("DEM","TWI", "NDRE.M")],
   dat$VW,
   method="rf",
   trControl = ctrl)
model_bnndm
}
}
