% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnndm_subset.R
\name{bnndm_cl}
\alias{bnndm_cl}
\title{NN distance based block CV}
\usage{
bnndm_cl(
  x = samplepoints,
  modeldomain = predictors,
  samplesize = 1000,
  sampling = "regular",
  w = 60,
  n_steps = 10,
  maxn = 10,
  cl = NULL
)
}
\arguments{
\item{x}{A sf object containing the training points.}

\item{modeldomain}{A raster or terra object containing the predictor variables.}

\item{samplesize}{Optional. The number of samplepoints generated in the modeldomain to describe the prediction space.
If NULL (default), the samplesize equals the number of cells in the predictor raster divided by ten.}

\item{sampling}{The type in which the samplepoints are generated. The default value is "Fibonacci", which is recommended for global studies.}

\item{w}{The window size. By default 60.}

\item{n_steps}{The number of steps until termination.}

\item{maxn}{The maximal number of clusters to be fitted. By default 10.}

\item{cl}{A cluster object.}
}
\description{
This function estimates the optimal cross-validation folds.
Therefore the function calculates the nearest neighbor distance
between training locations and prediction locations, as well as
between different cross-validation folds.
The resulting numerical vectors are then used to calculate
density distributions. The magnitude of difference between
the prediction to sample nearest neighbor density distribution to
that of different cross-validation folds is quantified using
the Wasserstein test. The cross-validation folds yielding
the lowest magnitude of distance towards the sample to prediction
nearest neighbor density distribution is then chosen as the final
split.
The function returns the density distribution plots,
the training points with assigned cross-validation folds,
the grid that was used to split the data, as well as
a caret object inheriting the folds.
}
\details{
The nearest neighbor distance distribution between training and prediction locations,
as well as between cross-validation folds are calculated. Based on that, the
cross-validation split yielding the lowest distance between those distributions is chosen.
}
